Restructured floating controls.
