--- src/components/Components/FloatingInfoButton.css
@@ -1,24 +1,0 @@
-/* FloatingButton.css */
-.floating-button-container {
-  position: absolute;
-  top: 80px;
-  right: 25px;
-  z-index: 999;
-}
-
-.floating-button {
-  border: none;
-  border-radius: 50%;
-  padding: 10px;
-  font-size: 20px;
-  cursor: pointer;
-}
-
-.floating-window {
-  position: fixed;
-  top: 120px; /* Adjust as needed */
-  right: 20px;
-  border: 1px solid;
-  padding: 10px;
-  z-index: 998;
-}
